Struct disjoint_set::DisjointSet

pub struct DisjointSet<T> {
    // some fields omitted
}

Struct that represents the Disjoint-Set data structure.

Methods

impl<T> DisjointSet<T> where T: Eq + PartialEq + Hash<Hasher> + Clone

fn new() -> DisjointSet<T>

fn make_set(&mut self, value: T)

Makes a singleton set of the value inside the DisjointSet.

fn find(&mut self, value: T) -> Option<T>

Finds the value of the root of the set that the value belongs to and performs path compression on the visited nodes.

Returns None if the value is not in the DisjointSet.

fn union(&mut self, value_one: T, value_two: T) -> Option<T>

Unions the two sets that each value belongs to using union by rank.

Returns None if one of the values does not exist in the DisjointSet.